requirements.txt
gradio==4.26.0
onnxruntime==1.17.3
numpy==1.26.4
Pillow==10.3.0
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
